Try fix windows issue (int64 vs int32)

+++ allensdk/brain_observatory/behavior/behavior_project_cache/tables/util/prior_exposure_processing.py
@@ -1,5 +1,6 @@
 from typing import Optional
 
+import numpy as np
 import pandas as pd
 
 from allensdk.brain_observatory.behavior.project_apis.data_io import \
@@ -173,7 +174,7 @@ def __get_prior_exposure_count(df: pd.DataFrame, to: pd.Series,
         df['to'] = to
 
         def cumsum(x):
-            return x.cumsum().shift(fill_value=0)
+            return x.cumsum().shift(fill_value=0).astype(np.int)
 
         counts = df.groupby(['mouse_id'])['to'].apply(cumsum)
         counts.name = None
